Calibration

+++ vision.c
@@ -6,12 +6,12 @@
 #define BLUEGREEN_MUL 3
 #define AREA_MIN 1000
 
-static int hueMin    = -0.3f;
-static int hueMax    =  0.1f;
-static int brightMin =  0.f;
-static int brightMax =  1.f;
-static int satMin    =  0.f;
-static int satMax    =  1.f;
+static float hueMin    = -0.3f;
+static float hueMax    =  0.1f;
+static float brightMin =  0.f;
+static float brightMax =  1.f;
+static float satMin    =  0.f;
+static float satMax    =  1.f;
 
 #define min(x, y) ((x) < (y) ? (x) : (y))
